refactor: 💡 split up test suite
